Store GeoJSON schools data in the DOM
Originally, to keep the home page lightweight, no schools data was in
the HTML. It was loaded onto the map via an AJAX request for a file,
`/schools.geojson`.

This is slightly problematic because it means that
1) Users not using javascript have no path to access the schools pages.
2) Search engines may have a hard time finding the schools pages to
index them.

Recently, a "Browse" dropdown was added to see a list of all schools.
Writing that list into the HTML when generating the site makes sense,
and solves both of the above problems.

Additionally, since most of the weight of the `browse` list in the HTML
is the HTML itself, and the required additional info is minimal, this
commit adds all of the data previously pulled via `schools.geojson`
into the home page HTML.

After GZIP, the extra weight on the home page is small, since the
markup and content are very repetitive. The resulting experience is
faster, since no AJAX is necessary after loading the page any more,
just parsing a bit of the DOM.

The actual dom parse step saves the data separately from the DOM, in
GeoJSON format, so working with it on the map is simple.