Be more backbone-y and fix a bad mem/event leak
Listening to events on views' models/collections is now binded with
`this.listenTo` instead of `model.on`. This stopped one big event leak.

Updating the displayed schools list based on the sorted results of the
fuzzy search used to leak dom nodes with attached events like crazy.
now:

 * Each list item's view is created exactly once, when it is added to
   the collection
 * a mapping from each school's `cid` to the list item view is
   maintained by the container's view, so that inserting and shuffling
   the order of the list just moves existing DOM nodes around.

And some other tweaks to the new backbone code.